chore: move code related to plugin to a dedicated file
